rename: api 폴더 이름 변경 (#18)
- 기존 server 이름에서 api 이름으로 수정
